+++ vector/src/vector_types/general/statistics.rs
@@ -366,7 +366,7 @@ where
         let array = self.data.to_slice();
         let chunks = Chunk::get_chunked_results(
             Complexity::Small,
-            self.multicore_settings,
+            &self.multicore_settings,
             &array[0..data_length],
             1,
             (),
@@ -407,7 +407,7 @@ where
         let array = self.data.to_slice();
         let chunks = Chunk::get_chunked_results(
             Complexity::Small,
-            self.multicore_settings,
+            &self.multicore_settings,
             &array[0..data_length],
             1,
             len,
@@ -442,7 +442,7 @@ where
         let sum = {
             let chunks = Chunk::get_chunked_results(
                 Complexity::Small,
-                self.multicore_settings,
+                &self.multicore_settings,
                 partition.center(array),
                 Reg::LEN,
                 (),
@@ -472,7 +472,7 @@ where
         let sum =  {
             let chunks = Chunk::get_chunked_results(
                 Complexity::Small,
-                self.multicore_settings,
+                &self.multicore_settings,
                 partition.center(array),
                 Reg::LEN,
                 (),
@@ -502,7 +502,7 @@ where
         let sum = {
             let chunks = Chunk::get_chunked_results(
                 Complexity::Small,
-                self.multicore_settings,
+                &self.multicore_settings,
                 partition.center(array),
                 Reg::LEN,
                 (),
@@ -532,7 +532,7 @@ where
         let sum = {
             let chunks = Chunk::get_chunked_results(
                 Complexity::Small,
-                self.multicore_settings,
+                &self.multicore_settings,
                 partition.center(array),
                 Reg::LEN,
                 (),
@@ -586,7 +586,7 @@ where
         let array = self.data.to_slice();
         let chunks = Chunk::get_chunked_results(
             Complexity::Small,
-            self.multicore_settings,
+            &self.multicore_settings,
             &array[0..data_length],
             2,
             (),
@@ -628,7 +628,7 @@ where
         let array = self.data.to_slice();
         let chunks = Chunk::get_chunked_results(
             Complexity::Small,
-            self.multicore_settings,
+            &self.multicore_settings,
             &array[0..data_length],
             2,
             len,